refactor: return @can in Tasks

+++ resources/views/tasks/index.blade.php
@@ -68,16 +68,20 @@
                             <td>{{ $task->created_at->format('d.m.Y') }}</td>
                             @auth
                                 <td>
-                                    <a href="{{ route('tasks.destroy', $task->id) }}"
-                                       data-confirm="{{ __('tasks.Confirm deletion') }}" data-method="delete"
-                                       rel="nofollow"
-                                       class="btn btn-danger btn-sm text-red-600 hover:text-red-900">
-                                        {{ __('tasks.Delete') }}
-                                    </a>
-                                    <a href="{{ route('tasks.edit', $task->id) }}"
-                                       class="text-blue-600 hover:text-blue-900">
-                                        {{ __('tasks.Edit') }}
-                                    </a>
+                                    @can('delete', $task)
+                                          <a href="{{ route('tasks.destroy', $task->id) }}"
+                                             data-confirm="{{ __('tasks.Confirm deletion') }}" data-method="delete"
+                                             rel="nofollow"
+                                             class="btn btn-danger btn-sm text-red-600 hover:text-red-900">
+                                              {{ __('tasks.Delete') }}
+                                          </a>
+                                    @endcan
+                                    @can('update', $task)
+                                        <a href="{{ route('tasks.edit', $task->id) }}"
+                                           class="text-blue-600 hover:text-blue-900">
+                                            {{ __('tasks.Edit') }}
+                                        </a>
+                                    @endcan
                                 </td>
                             @endauth
                         </tr>
